game db functions now passing all tests (getGameEntries, createGameEntry, clearGameEntries

+++ server/db/snippet-db.js
@@ -5,6 +5,9 @@ const {toAscii, toChar} = require('./betweenASCIIValues')
 const { pool } = require('./pool.js')
 
 
+const missingRequiredParams = (name, obj) => {
+    return {success: false, error: `missing required params in ${name}: ${obj}`};
+}
 
 const getSnippetByLengthAndType = async (length, type) => {
     const result = await getSnippetByType(type)
@@ -12,7 +15,8 @@ const getSnippetByLengthAndType = async (length, type) => {
 }
 
 const getSnippetByType = async (type) => {
-    if (!type || typeof type !== "string") return [];
+    if (!type || typeof type !== "string") 
+        return missingRequiredParams("type", type);
     try {
         const connection = await pool.getConnection();
         const query = `
@@ -57,11 +61,13 @@ const getSnippetByType = async (type) => {
 
 const createSnippet = async (snippet) => {
 
-    let connection;
     const { id, type, length, data } = snippet;
 
+    if (!(id && type && length && data)) 
+        return missingRequiredParams("snippet", snippet);
+
     try {
-        connection = await pool.getConnection();
+        const connection = await pool.getConnection();
         const recordQuery = "INSERT INTO snippet_records (id, snippet_type, snippet_length) VALUES (?, ?, ?);";
         await connection.query(recordQuery, [id, type, length]);
         const preparedValues = [];
@@ -97,6 +103,7 @@ const createSnippet = async (snippet) => {
 };
 
 const deleteSnippetByID = async (id) => {
+    if (!id) return missingRequiredParams("id", id);
     const connection = await pool.getConnection();
     try {
         const query1 = 'DELETE FROM snippet_records WHERE id = ?';
@@ -116,6 +123,7 @@ const deleteSnippetByID = async (id) => {
 };
 
 const getSnippetByLength = async (length) => {
+    if (!length) return missingRequiredParams("length", length);
     try {
         const connection = await pool.getConnection();
         const query = `
@@ -126,7 +134,7 @@ const getSnippetByLength = async (length) => {
             ORDER BY data.line_index ASC;
         `;
         const result = await connection.query(query, [length]);
-        connection.release();
+        await connection.release();
         const characterConvertedData = result[0].map(line_data => {
             return {...line_data, line_text: toChar(JSON.parse(line_data.line_text)).join('')}
         });
@@ -162,6 +170,7 @@ const getSnippetByLength = async (length) => {
 
 
 const getSnippetByID = async (id) => {
+    if (!id) return missingRequiredParams("id", id);
     const connection = await pool.getConnection();
     const query = `
             SELECT rec.id, rec.snippet_type, rec.snippet_length, data.line_index, data.line_text 
@@ -185,20 +194,15 @@ const getSnippetByID = async (id) => {
             length: length,
             data: processedSnippetText
         };
-        connection.release();
+        await connection.release();
         return processedResult;
     } catch (error) {
         return {};
         console.error(error);
-    } finally {
-        connection.release();
-    }
+    } 
 };
 
-const getPool = () => pool;
 
-
-const closePool = async () => { pool.end(); }
 
 
 module.exports = {
@@ -208,6 +212,4 @@ module.exports = {
     createSnippet, 
     deleteSnippetByID,
     getSnippetByLengthAndType,
-    getPool,
-    closePool
 }
